requirements-colab.txt
# Keep it lean; Colab already has many packages
albumentations
opencv-python
pandas
pydicom
Pillow
tqdm
PyYAML
scikit-image
# segmentation-models-pytorch  # uncomment only if you need SMP here
